Provide additional clarity about plugins
When reading the descriptions about the plugins, it is not readily apparent if the plugin comes with the base install of Logstash or needs to be installed separately.  By simply stating if it is or is not included in the default install, users who are investigating new capabilities can immediately see if they need to perform additional steps to make use of the filters they would like to implement.